+++ Core/CleanArchSample.Application/Features/Brands/Queries/GetAllBrands/GetAllBrandsQueryMapperProfile.cs
@@ -1,0 +1,12 @@
+﻿using AutoMapper;
+using CleanArchSample.Domain.Entities;
+
+namespace CleanArchSample.Application.Features.Brands.Queries.GetAllBrands;
+
+public class GetAllBrandsQueryMapperProfile : Profile
+{
+    public GetAllBrandsQueryMapperProfile()
+    {
+        CreateMap<Brand, GetAllBrandsQueryResponse>();
+    }
+}
