+++ Core/CleanArchSample.Application/Features/Brands/Queries/GetAllBrands/GetAllBrandsQueryRequest.cs
@@ -1,0 +1,7 @@
+﻿using MediatR;
+
+namespace CleanArchSample.Application.Features.Brands.Queries.GetAllBrands;
+
+public class GetAllBrandsQueryRequest : IRequest<IReadOnlyList<GetAllBrandsQueryResponse>>
+{
+}
