+++ Core/CleanArchSample.Application/Features/Brands/Queries/GetAllBrands/GetAllBrandsQueryResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace CleanArchSample.Application.Features.Brands.Queries.GetAllBrands;
+
+public class GetAllBrandsQueryResponse
+{
+    public string Name { get; set; }
+}
