--- Core/CleanArchSample.Application/Features/Products/Queries/GetAllBrands/GetAllBrandsQueryHandler.cs
@@ -1,24 +1,0 @@
-﻿using AutoMapper;
-using CleanArchSample.Application.Features.Common;
-using CleanArchSample.Application.Interfaces.UnitOfWorks;
-using CleanArchSample.Domain.Entities;
-using MediatR;
-
-namespace CleanArchSample.Application.Features.Products.Queries.GetAllBrands
-{
-    public class GetAllBrandsQueryHandler : CqrsHandlerBase, IRequestHandler<GetAllBrandsQueryRequest, IReadOnlyList<GetAllBrandsQueryResponse>>
-    {
-        public GetAllBrandsQueryHandler(IUnitOfWork unitOfWork, IMapper mapper) : base(unitOfWork, mapper)
-        {
-        }
-        public async Task<IReadOnlyList<GetAllBrandsQueryResponse>> Handle(GetAllBrandsQueryRequest request, CancellationToken cancellationToken)
-        {
-            var brands = await UnitOfWork.GetReadRepository<Brand>()
-                .ToListAsync(cancellationToken: cancellationToken);
-
-            var response = Mapper.Map<IReadOnlyList<Brand>, IReadOnlyList<GetAllBrandsQueryResponse>>(brands);
-
-            return response;
-        }
-    }
-}
