--- Core/CleanArchSample.Application/Features/Products/Queries/GetAllBrands/GetAllBrandsQueryRequest.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace CleanArchSample.Application.Features.Products.Queries.GetAllBrands
-{
-    public class GetAllBrandsQueryRequest : IRequest<IReadOnlyList<GetAllBrandsQueryResponse>>
-    {
-    }
-}
