--- Core/CleanArchSample.Application/Features/Products/Queries/GetAllBrands/GetAllBrandsQueryResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace CleanArchSample.Application.Features.Products.Queries.GetAllBrands
-{
-    public class GetAllBrandsQueryResponse
-    {
-        public string Name { get; set; }
-    }
-}
