+++ Core/CleanArchSample.Application/Features/Products/Queries/GetAllProducts/Dtos/BrandDto.cs
@@ -1,0 +1,6 @@
+﻿namespace CleanArchSample.Application.Features.Products.Queries.GetAllProducts.Dtos;
+
+public class BrandDto
+{
+    public string Name { get; set; }
+}
