+++ Core/CleanArchSample.Application/Features/Products/Queries/GetAllProducts/GetAllProductsQueryResponse.cs
@@ -1,4 +1,6 @@
-﻿namespace CleanArchSample.Application.Features.Products.Queries.GetAllProducts
+﻿using CleanArchSample.Application.Features.Products.Queries.GetAllProducts.Dtos;
+
+namespace CleanArchSample.Application.Features.Products.Queries.GetAllProducts
 {
     public class GetAllProductsQueryResponse
     {
@@ -6,5 +8,6 @@
         public string Description { get; set; }
         public decimal Price { get; set; }
         public decimal Discount { get; set; }
+        public BrandDto Brand { get; set; }
     }
 }
